+++ database/seeds/DatabaseTruncater.php
@@ -13,6 +13,7 @@ class DatabaseTruncater extends Seeder
         DB::statement('TRUNCATE tokens CASCADE');
         DB::statement('TRUNCATE notifications CASCADE');
         DB::statement('TRUNCATE users CASCADE');
+        DB::statement('TRUNCATE recruiters CASCADE');
         DB::statement('ALTER TABLE tokens ENABLE TRIGGER ALL;');
         DB::statement('COMMIT;');
     }